[IMP] Added tests_rest_cases
